+++ app/Listing/Application/UseCases/DeleteListingMedia/DeleteListingMediaOutput.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Listing\Application\UseCases\DeleteListingMedia;
+
+use App\Shared\Domain\DTO\Output;
+
+class DeleteListingMediaOutput extends Output
+{
+    /**
+     * @param array<string, mixed> $item
+     */
+    public function __construct(
+        public readonly array $item,
+    ) {}
+}